Refactor: rename movements to controller
- Rename movements plugin to controller
- Add Health component and HPBar marker
- Introduce damage system and HP bar integration
- Replace HUD marker with HudMap and wire HP bars
- Update messages to EntityMoved for movement events

+++ src/components/player.rs
@@ -5,5 +5,5 @@ use bevy_ecs_tiled::prelude::*;
 #[derive(Component, Reflect, Default)]
 #[reflect(Component, Default)]
 pub struct Player {
-    pub player_id: u32,
+    pub player_id: usize,
 }
